Refactor the structure of the config manager and remove unused data managers
